28-07-2025

+++ WebApplication1/Models/DTO/AddBomDto.cs
@@ -11,11 +11,11 @@ namespace WebApplication1.Models.DTO
         public int bomid { get; set; }
    
         public int itemid { get; set; }
-        public double bomqty { get; set; }
+        public decimal bomqty { get; set; }
 
      
         public int bomuomid { get; set; }
-        public double price { get; set; }
+        public decimal price { get; set; }
       
         public int prodstageid { get; set; }
         public DateTime RequiredDate { get; set; }
