+++ WebApplication1/Models/DTO/AddPrdetails.cs
@@ -17,7 +17,7 @@ namespace WebApplication1.Models.DTO
         public int bomid { get; set; }
 
 
-        public float prqty { get; set; }
+        public decimal prqty { get; set; }
         public int  pruomid { get; set; }
 
     }
