+++ WebApplication1/Models/DTO/GroupedItem.cs
@@ -4,8 +4,8 @@
     {
         public int orderid { get; set; }
         public int PoItemId { get; set; }
-            public double TotalQuantity { get; set; }
-            public double UnitPrice { get; set; }
+            public decimal TotalQuantity { get; set; }
+            public decimal UnitPrice { get; set; }
             public string Make { get; set; }
 
         public int prtblid { get; set; }
