+++ WebApplication1/Models/Domain/PRDetails.cs
@@ -11,9 +11,9 @@ namespace WebApplication1.Models.Domain
         public int pritemid { get; set; }
         public Bom Bom { get; set; }
         public int bomid { get; set; }
-        public float  pocreatedqty { get; set; } = 0;
+        public decimal  pocreatedqty { get; set; } = 0;
 
-        public float prqty { get; set; }
+        public decimal prqty { get; set; }
         public ICollection<Purchasedetails> Purchasedetails { get; set; }
         public ICollection<PRPO> PRPOs { get; set; } = new List<PRPO>();
         public UOM  UOM { get; set; }
